+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [DEV]
 
 * Throw when calling match? with times-to-try > 1 and a value for actual (should be a step) [#116](https://github.com/nubank/state-flow/pull/116)
+* Add `state-flow.core/fmap` [#117](https://github.com/nubank/state-flow/pull/117)
 
 ## [3.0.0]